Rename support.versionXX projects to support.web.versionXX
