+++ dev/com.ibm.websphere.appserver.springBoot-2.0/bnd.bnd
@@ -13,5 +13,5 @@
 -nobundles=true
 
 -dependson: \
-	com.ibm.ws.springboot.support.version20;version=latest
+	com.ibm.ws.springboot.support.web.version20;version=latest
