--- dev/com.ibm.ws.springboot.support.version15/bnd/springBoot15Starter.bnd
@@ -1,3 +1,0 @@
-Bundle-SymbolicName: springBoot15Starter
-Private-Package: com.ibm.ws.springboot.support.version15.container.*
-Include-Resource: resources/container
