--- dev/com.ibm.ws.springboot.support.version15/resources/container/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.ibm.ws.springboot.support.version15.container.LibertyServletContainerConfiguration
