--- dev/com.ibm.ws.springboot.support.version20/resources/container/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.ibm.ws.springboot.support.version20.container.LibertyServletContainerConfiguration
