--- dev/com.ibm.ws.springboot.support.version20/src/com/ibm/ws/springboot/support/version20/container/LibertyServletContainerConfiguration.java
@@ -1,38 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2018 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *     IBM Corporation - initial API and implementation
- *******************************************************************************/
-package com.ibm.ws.springboot.support.version20.container;
-
-import org.springframework.boot.autoconfigure.AutoConfigureOrder;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
-import org.springframework.boot.autoconfigure.condition.SearchStrategy;
-import org.springframework.boot.autoconfigure.web.servlet.ServletWebServerFactoryAutoConfiguration.BeanPostProcessorsRegistrar;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.boot.web.servlet.server.ServletWebServerFactory;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-import org.springframework.core.Ordered;
-
-@AutoConfigureOrder(Ordered.HIGHEST_PRECEDENCE)
-@Configuration
-@EnableConfigurationProperties
-@ConditionalOnWebApplication
-@Import(BeanPostProcessorsRegistrar.class)
-public class LibertyServletContainerConfiguration {
-
-    @Bean
-    @ConditionalOnMissingBean(value = ServletWebServerFactory.class, search = SearchStrategy.CURRENT)
-    public LibertyServletContainerFactory libertyEmbeddedServletContainerFactory() {
-        return new LibertyServletContainerFactory();
-    }
-
-}
