+++ dev/com.ibm.ws.springboot.support.web.version15/bnd/springBoot15Web.bnd
@@ -1,0 +1,3 @@
+Bundle-SymbolicName: springBoot15Web
+Private-Package: com.ibm.ws.springboot.support.web.version15.container.*
+Include-Resource: resources/container
