+++ dev/com.ibm.ws.springboot.support.web.version15/src/com/ibm/ws/springboot/support/web/version15/container/LibertyServletContainerFactory.java
@@ -1,0 +1,41 @@
+/*******************************************************************************
+ * Copyright (c) 2018 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+package com.ibm.ws.springboot.support.web.version15.container;
+
+import org.springframework.beans.BeansException;
+import org.springframework.boot.context.embedded.AbstractEmbeddedServletContainerFactory;
+import org.springframework.boot.context.embedded.EmbeddedServletContainer;
+import org.springframework.boot.web.servlet.ServletContextInitializer;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.ApplicationContextAware;
+
+/**
+ *
+ */
+public class LibertyServletContainerFactory extends AbstractEmbeddedServletContainerFactory implements ApplicationContextAware {
+
+    ApplicationContext context;
+
+    @Override
+    public EmbeddedServletContainer getEmbeddedServletContainer(ServletContextInitializer... initializers) {
+        return new LibertyServletContainer(this, mergeInitializers(initializers));
+    }
+
+    @Override
+    public void setApplicationContext(ApplicationContext context) throws BeansException {
+        this.context = context;
+    }
+
+    String getApplicationID() {
+        return context.getId();
+    }
+
+}
