+++ dev/com.ibm.ws.springboot.support.web.version15/src/com/ibm/ws/springboot/support/web/version15/osgi/SpringBoot15Support.java
@@ -1,0 +1,23 @@
+/*******************************************************************************
+ * Copyright (c) 2018 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+package com.ibm.ws.springboot.support.web.version15.osgi;
+
+import org.osgi.service.component.annotations.Component;
+
+import com.ibm.ws.app.manager.springboot.support.SpringBootSupport;
+
+/**
+ *
+ */
+@Component(service = SpringBootSupport.class)
+public class SpringBoot15Support extends SpringBootSupport {
+
+}
