+++ dev/com.ibm.ws.springboot.support.web.version20/bnd/springBoot20Web.bnd
@@ -1,0 +1,3 @@
+Bundle-SymbolicName: springBoot20Web
+Private-Package: com.ibm.ws.springboot.support.web.version20.container.*
+Include-Resource: resources/container
