+++ dev/com.ibm.ws.springboot.support.web.version20/resources/container/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.ibm.ws.springboot.support.web.version20.container.LibertyServletContainerConfiguration
